split table file
