+++ packages/runtime/example/table/dynamic.html
@@ -103,7 +103,7 @@
               type: 'chakra_ui/v1/table',
               properties: {
                 size: 'sm',
-                data: '{{ fetchUser.query.data }}',
+                data: '{{ fetchUser.fetch.data }}',
                 majorKey: 'id',
                 rowsPerPage: 5,
                 columns: [
