+++ packages/runtime/src/components/chakra-ui/Table/index.ts
@@ -1,0 +1,48 @@
+import { createComponent } from '@meta-ui/core';
+import { TableImpl } from './Table';
+import {
+  ColumnsPropertySchema,
+  DataPropertySchema,
+  MajorKeyPropertySchema,
+  RowsPerPagePropertySchema,
+  TableStateSchema,
+  TableSizePropertySchema,
+} from './TableTypes';
+
+export default {
+  ...createComponent({
+    version: 'chakra_ui/v1',
+    metadata: {
+      name: 'table',
+      description: 'chakra-ui table',
+    },
+    spec: {
+      properties: [
+        {
+          name: 'data',
+          ...DataPropertySchema,
+        },
+        {
+          name: 'marjorKey',
+          ...MajorKeyPropertySchema,
+        },
+        {
+          name: 'rowsPerPage',
+          ...RowsPerPagePropertySchema,
+        },
+        {
+          name: 'size',
+          ...TableSizePropertySchema,
+        },
+        {
+          name: 'columns',
+          ...ColumnsPropertySchema,
+        },
+      ],
+      acceptTraits: [],
+      state: TableStateSchema,
+      methods: [],
+    },
+  }),
+  impl: TableImpl,
+};
