+++ packages/runtime/src/registry.tsx
@@ -16,7 +16,7 @@ import CoreDummy from './components/core/Dummy';
 import ChakraUIRoot from './components/chakra-ui/Root';
 import ChakraUIButton from './components/chakra-ui/Button';
 import ChakraUITabs from './components/chakra-ui/Tabs';
-import ChakraUITable from './components/chakra-ui/Table/Table';
+import ChakraUITable from './components/chakra-ui/Table';
 import ChakraUIInput from './components/chakra-ui/Input';
 import ChakraUIBox from './components/chakra-ui/Box';
 import ChakraUIKbd from './components/chakra-ui/Kbd';
